:skull: Add a way to delete the phisical mail folder as well

+++ src/services/MailUserService.php
@@ -74,8 +74,15 @@ class MailUserService {
   }
 
   public function deleteMail($data) {
-    $this->virtualuser->destroy($data[0]);
-    return true;
+    $mail_id = $data[0];
+    $mail_username = $data[1];
+    if (filter_var($mail_id, FILTER_VALIDATE_INT && !empty($mail_username))) {
+      $this->virtualuser->destroy($data[0]);
+      shell_exec('deletemailbox ' . escapeshellarg($mail_username));
+      return true;
+    } else {
+      return false;
+    }
   }
 
   public function isUniq($data) {
